+++ src/cougardevice.h
@@ -39,6 +39,9 @@ const int cCougarInterfaceBulkIn  = 4;
 const int cCougarEndpointBulkOut = 4;
 const int cCougarEndpointBulkIn  = 5 | 0x80;
 
+// Util only supports extracting and updating a single fw version at this time.
+const std::string cSupportedFirmwareVersion{ "3.00.6 revB" };
+
 //////////////////////////////////////////////////////////////////////
 // Cougar Options bitflags
 //////////////////////////////////////////////////////////////////////